# src/typescript/coreapis-sdk/models/invitation-public.ts
/* tslint:disable */
/* eslint-disable */
/**
 * FastAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import { Organization } from './organization';
import { User } from './user';
 /**
 * Public invitation info (without sensitive data like full token in some contexts)
 *
 * @export
 * @interface InvitationPublic
 */
export interface InvitationPublic {

    /**
     * @type {string}
     * @memberof InvitationPublic
     */
    id: string;

    /**
     * @type {string}
     * @memberof InvitationPublic
     */
    email: string;

    /**
     * @type {string}
     * @memberof InvitationPublic
     */
    role: string;

    /**
     * @type {string}
     * @memberof InvitationPublic
     */
    status: string;

    /**
     * @type {Organization}
     * @memberof InvitationPublic
     */
    organization: Organization;

    /**
     * @type {User}
     * @memberof InvitationPublic
     */
    inviter: User;

    /**
     * @type {number}
     * @memberof InvitationPublic
     */
    expiresAt: number;

    /**
     * @type {number}
     * @memberof InvitationPublic
     */
    creationTime: number;
}