# src/typescript/coreapis-sdk/models/invitation.ts
/* tslint:disable */
/* eslint-disable */
/**
 * FastAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import { Organization } from './organization';
 /**
 * Complete invitation model with all fields
 *
 * @export
 * @interface Invitation
 */
export interface Invitation {

    /**
     * @type {string}
     * @memberof Invitation
     */
    id: string;

    /**
     * @type {string}
     * @memberof Invitation
     */
    email: string;

    /**
     * @type {string}
     * @memberof Invitation
     */
    orgId: string;

    /**
     * @type {Organization}
     * @memberof Invitation
     */
    organization?: Organization;

    /**
     * @type {string}
     * @memberof Invitation
     */
    inviterId: string;

    /**
     * @type {string}
     * @memberof Invitation
     */
    role: string;

    /**
     * @type {string}
     * @memberof Invitation
     */
    token: string;

    /**
     * @type {string}
     * @memberof Invitation
     */
    status: string;

    /**
     * @type {number}
     * @memberof Invitation
     */
    expiresAt: number;

    /**
     * @type {number}
     * @memberof Invitation
     */
    acceptedAt?: number | null;

    /**
     * @type {boolean}
     * @memberof Invitation
     */
    deleted: boolean;

    /**
     * @type {number}
     * @memberof Invitation
     */
    creationTime: number;

    /**
     * @type {number}
     * @memberof Invitation
     */
    lastUpdateTime: number;
}