# src/typescript/coreapis-sdk/openapi.yaml
openapi: 3.0.1
info:
  title: FastAPI
  version: 0.1.0
paths:
...
  /v1/invitations:
    post:
      tags:
      - Invitations
      summary: Create Invitation
      description: 'Create a new invitation. The payload must include orgId and
        the invitee joins that existing organization.


        Permissions:

        - Company Owner can invite managers/standard/agronomists

        - Agronomist can invite users into an existing managed organization

        - Company Manager can invite standard users'
      operationId: create_invitation
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvitationCreatePayload'
        required: true
      responses:
        '200':
          description: Invitation created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invitation'
        '401':
          description: Unauthorized to access this resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Unauthorized to access this resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - SecurityChecker: []
  /v1/invitations/validate:
    get:
      tags:
      - Invitations
      summary: Validate Invitation Token
      description: 'Validate an invitation token (public endpoint, no auth required)

        Returns invitation details if valid'
      operationId: validate_invitation_token
      parameters:
      - name: token
        in: query
        required: true
        schema:
          type: string
          description: Invitation token
          title: Token
        description: Invitation token
      responses:
        '200':
          description: Invitation validation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvitationValidateResponse'
        '401':
          description: Unauthorized to access this resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Unauthorized to access this resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/invitations/accept:
    post:
      tags:
      - Invitations
      summary: Accept Invitation
      description: 'Accept an invitation (user must be authenticated)

        Adds user to organization and assigns role'
      operationId: accept_invitation
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvitationAcceptPayload'
        required: true
      responses:
        '200':
          description: Invitation accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '401':
          description: Unauthorized to access this resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Unauthorized to access this resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - SecurityChecker: []
  /v1/invitations/decline:
    post:
      tags:
      - Invitations
      summary: Decline Invitation
      description: Decline an invitation (no auth required - uses token)
      operationId: decline_invitation
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvitationAcceptPayload'
        required: true
      responses:
        '200':
          description: Invitation declined
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '401':
          description: Unauthorized to access this resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Unauthorized to access this resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/invitations/byOrg:
    get:
      tags:
      - Invitations
      summary: List Organization Invitations
      description: 'List all invitations for an organization

        Filter by status: pending, accepted, declined, expired'
      operationId: list_organization_invitations
      security:
      - SecurityChecker: []
      parameters:
      - name: org_id
        in: query
        required: true
        schema:
          type: string
          description: Organization ID
          title: Org Id
        description: Organization ID
      - name: status
        in: query
        required: false
        schema:
          description: Filter by status
          default: pending
          title: Status
          type: string
          nullable: true
        description: Filter by status
      responses:
        '200':
          description: List of invitations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Invitation'
                title: Response List Organization Invitations
        '401':
          description: Unauthorized to access this resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Unauthorized to access this resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/invitations/me:
    get:
      tags:
      - Invitations
      summary: List My Invitations
      description: 'List all pending invitations for the authenticated user

        Matches by email from Keycloak profile'
      operationId: list_my_invitations
      responses:
        '200':
          description: List of pending invitations
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Invitation'
                type: array
                title: Response List My Invitations
        '401':
          description: Unauthorized to access this resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Unauthorized to access this resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - SecurityChecker: []
  /v1/invitations/{invitation_id}:
    delete:
      tags:
      - Invitations
      summary: Revoke Invitation
      description: Revoke (soft delete) a pending invitation
      operationId: revoke_invitation
      security:
      - SecurityChecker: []
      parameters:
      - name: invitation_id
        in: path
        required: true
        schema:
          type: string
          title: Invitation Id
      responses:
        '200':
          description: Invitation revoked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '401':
          description: Unauthorized to access this resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Unauthorized to access this resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/invitations/{invitation_id}/resend:
    post:
      tags:
      - Invitations
      summary: Resend Invitation
      description: Resend invitation email (generates new token)
      operationId: resend_invitation
      security:
      - SecurityChecker: []
      parameters:
      - name: invitation_id
        in: path
        required: true
        schema:
          type: string
          title: Invitation Id
      responses:
        '200':
          description: Invitation resent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '401':
          description: Unauthorized to access this resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Unauthorized to access this resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    AccountTypeEnum:
      type: string
      enum:
      - Admin
      - Agronomist
      - Standard
      title: AccountTypeEnum
...
    Contacts:
      properties:
        email:
          type: string
          title: Email
        phone:
          type: string
          title: Phone
      type: object
      required:
      - email
      - phone
      title: Contacts
...
    ErrorResponse:
      properties:
        status:
          type: integer
          title: Status
        detail:
          anyOf:
          - type: string
          - type: object
          title: Detail
      type: object
      required:
      - status
      - detail
      title: ErrorResponse
...
    Invitation:
      properties:
        id:
          type: string
          title: Id
        email:
          type: string
          title: Email
        orgId:
          title: Orgid
          type: string
        organization:
          $ref: '#/components/schemas/Organization'
          nullable: true
        inviterId:
          type: string
          title: Inviterid
        role:
          type: string
          title: Role
        token:
          type: string
          title: Token
        status:
          type: string
          title: Status
        expiresAt:
          type: integer
          title: Expiresat
        acceptedAt:
          title: Acceptedat
          type: integer
          nullable: true
        deleted:
          type: boolean
          title: Deleted
        creationTime:
          type: integer
          title: Creationtime
        lastUpdateTime:
          type: integer
          title: Lastupdatetime
      type: object
      required:
      - id
      - email
      - orgId
      - inviterId
      - role
      - token
      - status
      - expiresAt
      - deleted
      - creationTime
      - lastUpdateTime
      title: Invitation
      description: Complete invitation model with all fields
    InvitationAcceptPayload:
      properties:
        token:
          type: string
          title: Token
      type: object
      required:
      - token
      title: InvitationAcceptPayload
      description: Payload for accepting invitation
    InvitationCreatePayload:
      properties:
        email:
          type: string
          format: email
          title: Email
        role:
          type: string
          title: Role
        orgId:
          title: Orgid
          type: string
      type: object
      required:
      - email
      - role
      - orgId
      title: InvitationCreatePayload
      description: Payload for creating an invitation
    InvitationPublic:
      properties:
        id:
          type: string
          title: Id
        email:
          type: string
          title: Email
        role:
          type: string
          title: Role
        status:
          type: string
          title: Status
        organization:
          $ref: '#/components/schemas/Organization'
        inviter:
          $ref: '#/components/schemas/User'
        expiresAt:
          type: integer
          title: Expiresat
        creationTime:
          type: integer
          title: Creationtime
      type: object
      required:
      - id
      - email
      - role
      - status
      - organization
      - inviter
      - expiresAt
      - creationTime
      title: InvitationPublic
      description: Public invitation info (without sensitive data like full token in some contexts)
    InvitationValidateResponse:
      properties:
        valid:
          type: boolean
          title: Valid
        invitation:
          $ref: '#/components/schemas/InvitationPublic'
          nullable: true
        error:
          title: Error
          type: string
          nullable: true
      type: object
      required:
      - valid
      title: InvitationValidateResponse
      description: Response for token validation
...
    Organization:
      properties:
        orgId:
          type: string
          title: Orgid
        name:
          type: string
          title: Name
        piva:
          type: string
          title: Piva
        logo:
          type: string
          title: Logo
        cover:
          type: string
          title: Cover
        contacts:
          $ref: '#/components/schemas/Contacts'
        creationTime:
          type: integer
          title: Creationtime
        lastUpdateTime:
          type: integer
          title: Lastupdatetime
      type: object
      required:
      - orgId
      - name
      - piva
      - logo
      - cover
      - contacts
      - creationTime
      - lastUpdateTime
      title: Organization
...
    StatusResponse:
      properties:
        status:
          type: integer
          title: Status
        message:
          type: string
          title: Message
      type: object
      required:
      - status
      - message
      title: StatusResponse
    User:
      properties:
        id:
          type: string
          title: Id
        firstName:
          type: string
          title: Firstname
        lastName:
          type: string
          title: Lastname
        email:
          type: string
          title: Email
        emailVerified:
          type: boolean
          title: Emailverified
        enabled:
          type: boolean
          title: Enabled
        accountType:
          $ref: '#/components/schemas/AccountTypeEnum'
        phone:
          type: string
          title: Phone
        piva:
          type: string
          title: Piva
        avatar:
          type: string
          title: Avatar
        organizations:
          items:
            $ref: '#/components/schemas/UserOrgananizationMembership'
          type: array
          title: Organizations
          default: []
        creationTime:
          type: integer
          title: Creationtime
      type: object
      required:
      - id
      - firstName
      - lastName
      - email
      - emailVerified
      - enabled
      - accountType
      - phone
      - piva
      - avatar
      - creationTime
      title: User
...
    UserOrgananizationMembership:
      properties:
        id:
          type: string
          title: Id
        name:
          type: string
          title: Name
        roles:
          items:
            type: string
          type: array
          title: Roles
      type: object
      required:
      - id
      - name
      - roles
      title: UserOrgananizationMembership
...
  securitySchemes:
    SecurityChecker:
      type: http
      scheme: bearer